@@ -1,53 +1,57 @@
 {
   "name": "auro-wallet",
-  "version": "2.4.4",
+  "version": "2.4.5",
   "private": true,
   "dependencies": {
-    "@ledgerhq/hw-transport-webusb": "6.29.4",
-    "@reduxjs/toolkit": "1.4.0",
+    "@ledgerhq/hw-transport-webhid": "6.30.9",
+    "@ledgerhq/devices": "8.7.0",
+    "@reduxjs/toolkit": "2.11.1",
     "@aurowallet/mina-provider": "1.0.12",
-    "axios": "1.10.0",
-    "axios-retry":"4.5.0",
+    "axios": "1.13.2",
+    "axios-retry": "4.5.0",
     "bignumber.js": "9.3.0",
-    "@scure/bip32":"1.7.0",
-    "@scure/bip39":"1.6.0",
-    "bs58check": "2.1.2",
+    "@scure/bip32": "1.7.0",
+    "@scure/bip39": "1.6.0",
+    "@scure/base": "^2.0.0",
     "classnames": "2.5.1",
     "webextension-polyfill": "0.12.0",
-    "i18next": "19.8.4",
-    "libsodium-wrappers": "0.7.9",
+    "i18next": "25.7.3",
+    "libsodium-wrappers-sumo": "0.7.15",
     "mina-signer": "3.1.0",
-    "@zondax/ledger-mina-js":"0.1.1",
-    "obs-store": "4.0.3",
-    "react": "17.0.1",
-    "react-circular-progressbar": "2.0.3",
-    "react-dom": "17.0.1",
-    "react-helmet": "6.1.0",
-    "react-i18next": "11.8.2",
-    "react-idle-timer": "4.2.5",
-    "react-redux": "7.2.2",
-    "react-router-dom": "5.1.2",
-    "react-scripts": "5.0.1",
-    "redux": "4.0.5",
+    "@zondax/ledger-mina-js": "0.1.1",
+    "zustand":"5.0.9",
+    "react": "19.2.3",
+    "react-dom": "19.2.3",
+    "react-i18next": "16.5.0",
+    "react-idle-timer": "5.7.2",
+    "react-redux": "9.2.0",
+    "react-router-dom": "7.11.0",
+    "react-circular-progressbar": "2.2.0",
     "buffer": "6.0.3",
     "uuid": "11.1.0",
-    "valid-url": "1.0.9",
     "qrcode.react": "4.2.0",
-    "styled-components":"6.1.19",
-    "node-forge":"1.3.1",
-    "o1js":"2.4.0",
-    "serialize-error":"12.0.0",
-    "mina-attestations":"0.5.0",
-    "@noble/hashes":"1.7.1",
-    "immer":"10.1.1"
+    "styled-components": "6.1.19",
+    "node-forge": "1.3.3",
+    "o1js": "2.4.0",
+    "serialize-error": "12.0.0",
+    "mina-attestations": "0.5.0",
+    "@noble/hashes": "1.7.1",
+    "immer": "10.1.1"
   },
   "scripts": {
-    "build": "export NODE_OPTIONS=--openssl-legacy-provider && rm -rf dist && webpack --mode production",
-    "postbuild": "rm -rf zip && node ./scripts/post-build.js",
-    "buildProd": "webpack --mode production",
-    "dev": "node ./scripts/post-dev.js && webpack --mode development --progress --watch",
+    "clean": "rm -rf dist dist-chrome dist-firefox zip",
+    "build:chrome": "rm -rf dist dist-chrome && webpack --mode production --env BROWSER=chrome && mv dist dist-chrome",
+    "build:firefox": "rm -rf dist dist-firefox && webpack --mode production --env BROWSER=firefox && mv dist dist-firefox",
+    "build": "npm run clean && npm run build:chrome && npm run build:firefox",
+    "postbuild": "node scripts/post-build.js",
+    "predev": "rm -rf dist && node scripts/pre-dev.js",
+    "dev": "webpack --mode development --env BROWSER=chrome   --watch",
+    "predev:ff": "rm -rf dist-firefox && BROWSER=firefox node scripts/pre-dev.js",
+    "dev:ff": "webpack --mode development --env BROWSER=firefox  --watch --output-path=dist-firefox",
     "test": "mocha --timeout 5000 --require ./test/babel-register.js 'test/unit/**/*.js'",
-    "cspell":"cspell \"src/**\""
+    "testa": "mocha --timeout 5000 --require ./test/babel-register.js 'test/unit/apiService.spec.js'",
+    "testb": "mocha --timeout 5000 --require ./test/babel-register.js 'test/unit/dappService.spec.js'",
+    "cspell": "cspell \"src/**\""
   },
   "eslintConfig": {
     "extends": [
