@@ -1,20 +1,16 @@
 import i18n from "i18next";
 import { useCallback } from "react";
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import CustomView from "../../component/CustomView";
 import styles from "./index.module.scss";
 
 
 const Security = ({ }) => {
-  const history = useHistory()
-  const onToSeedPage = useCallback(() => {
-    history.push('reveal_seed_page')
-  }, [])
+  const navigate = useNavigate()
   const onResetPwd = useCallback(() => {
-    history.push('reset_password')
+    navigate('/reset_password')
   }, [])
   return <CustomView title={i18n.t('security')} contentClassName={styles.container}>
-    <RowItem title={i18n.t('backupMnemonicPhrase')} onClickItem={onToSeedPage} />
     <RowItem title={i18n.t('changePassword')} onClickItem={onResetPwd} />
   </CustomView>
 }
