@@ -4,18 +4,15 @@ import i18n from "i18next";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { Trans } from "react-i18next";
 import { useDispatch, useSelector } from "react-redux";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import { getCurrencyPrice } from "../../../background/api";
 
-import { saveLocal } from "@/background/localStorage";
 import { NetworkID_MAP } from "@/constant/network";
-import { LOCAL_CACHE_KEYS } from "@/constant/storageKey";
 import useFetchAccountData from "@/hooks/useUpdateAccount";
 import Clock from "@/popup/component/Clock";
 import styled, { css } from "styled-components";
 import {
   DAPP_DISCONNECT_SITE,
-  DAPP_GET_CONNECT_STATUS,
   WALLET_GET_ALL_ACCOUNT,
 } from "../../../constant/msgTypes";
 import {
@@ -24,15 +21,15 @@ import {
   updateShouldRequest,
 } from "../../../reducers/accountReducer";
 import { setAccountInfo } from "../../../reducers/cache";
+import {
+  copyText,
+} from "../../../utils/browserUtils";
 import { sendMsg } from "../../../utils/commonMsg";
 import {
   addressSlice,
   getAmountForUI,
   isZekoNet,
 } from "../../../utils/utils";
-import {
-  copyText,
-} from "../../../utils/browserUtils";
 import { PopupModal } from "../../component/PopupModal";
 import Toast from "../../component/Toast";
 import NetworkSelect from "../Networks/NetworkSelect";
@@ -89,16 +86,16 @@ const StyledTipsSpecial = styled.span`
   color: #d65a5a;
 `;
 const Wallet = ({}) => {
-  const history = useHistory();
+  const navigate = useNavigate();
 
   const [watchModalStatus, setWatchModalStatus] = useState(false);
 
   const toSetting = useCallback(() => {
-    history.push("setting");
+    navigate("/setting");
   }, []);
 
   const toManagePage = useCallback(() => {
-    history.push("account_manage");
+    navigate("/account_manage");
   }, []);
 
   const onCloseWatchModal = useCallback(() => {
@@ -314,7 +311,7 @@ const WalletInfo = () => {
   const shouldRefresh = useSelector((state) => state.accountInfo.shouldRefresh);
 
   const dispatch = useDispatch();
-  const history = useHistory();
+  const navigate = useNavigate();
 
   const [currentAccount, setCurrentAccount] = useState(
     accountInfo.currentAccount
@@ -392,17 +389,17 @@ const WalletInfo = () => {
 
   const toAccountInfo = useCallback(() => {
     dispatch(setAccountInfo(currentAccount));
-    history.push("account_info");
+    navigate("/account_info");
   }, [currentAccount]);
 
   const toSend = useCallback(() => {
     setTokenModalStatus(true);
   }, []);
   const toReceive = useCallback(() => {
-    history.push("receive_page");
+    navigate("/receive_page");
   }, []);
   const toStaking = useCallback(() => {
-    history.push("staking");
+    navigate("/staking");
   }, []);
 
 
