@@ -69,31 +69,9 @@ export const WALLET_CHANGE_SEC_PASSWORD = "WALLET_CHANGE_SEC_PASSWORD"
 
 
 /**
- * Get the current private key
- */
-export const WALLET_GET_CURRENT_PRIVATE_KEY = "WALLET_GET_CURRENT_PRIVATE_KEY"
-
-
-/**
- * send transaction
- */
-export const WALLET_SEND_TRANSACTION = "WALLET_SEND_TRANSACTION"
-
-/**
- * send message transaction
- */
-export const WALLET_SEND_MESSAGE_TRANSACTION = "WALLET_SEND_MESSAGE_TRANSACTION"
-
-
-/**
  * sign fields
  */
 export const WALLET_SEND_FIELDS_MESSAGE_TRANSACTION = "WALLET_SEND_FIELDS_MESSAGE_TRANSACTION"
-
-/**
- * Send a stake transaction
- */
-export const WALLET_SEND_STAKE_TRANSACTION = "WALLET_SEND_STAKE_TRANSACTION"
 
 /**
  * Query stake status
@@ -104,12 +82,6 @@ export const WALLET_CHECK_TX_STATUS = "WALLET_CHECK_TX_STATUS"
  * Import ledger wallet
  */
 export const WALLET_IMPORT_LEDGER = "WALLET_IMPORT_LEDGER"
-
-/**
- * Import watch mode wallet
- */
-export const WALLET_IMPORT_WATCH_MODE = "WALLET_IMPORT_WATCH_MODE"
-
 
 /**
 * Import keystore wallet
@@ -214,11 +186,6 @@ export const DAPP_ACTION_GET_ACCOUNT = "DAPP_ACTION_GET_ACCOUNT"
 export const DAPP_ACTION_CLOSE_WINDOW = "DAPP_ACTION_CLOSE_WINDOW"
 
 /**
- * get current web connect status
- */
-export const DAPP_GET_CONNECT_STATUS = "DAPP_GET_CONNECT_STATUS"
-
-/**
  * disconnect web site
  */
 export const DAPP_DISCONNECT_SITE = "DAPP_DISCONNECT_SITE"
@@ -306,3 +273,31 @@ export const CredentialMsg = {
 }
 export const DAPP_ACTION_STORE_CREDENTIAL = "DAPP_ACTION_STORE_CREDENTIAL"
 export const DAPP_ACTION_REQUEST_PRESENTATION = "DAPP_ACTION_REQUEST_PRESENTATION"
+
+// ============================================
+// Multi-Wallet Message Types
+// ============================================
+
+/** Get all keyrings (wallet groups) sorted by creation time */
+export const WALLET_GET_KEYRINGS_LIST = "WALLET_GET_KEYRINGS_LIST"
+
+/** Add a new HD wallet with its own mnemonic */
+export const WALLET_ADD_HD_KEYRING = "WALLET_ADD_HD_KEYRING"
+
+/** Rename a keyring (wallet group) */
+export const WALLET_RENAME_KEYRING = "WALLET_RENAME_KEYRING"
+
+/** Get mnemonic for a specific HD keyring */
+export const WALLET_GET_KEYRING_MNEMONIC = "WALLET_GET_KEYRING_MNEMONIC"
+
+/** Delete a keyring (wallet group) */
+export const WALLET_DELETE_KEYRING = "WALLET_DELETE_KEYRING"
+
+/** Add account to a specific HD keyring */
+export const WALLET_ADD_ACCOUNT_TO_KEYRING = "WALLET_ADD_ACCOUNT_TO_KEYRING"
+
+/** Get current vault version (v1 or v2) */
+export const WALLET_GET_VAULT_VERSION = "WALLET_GET_VAULT_VERSION"
+
+/** Try to upgrade vault from v1 to v2 */
+export const WALLET_TRY_UPGRADE_VAULT = "WALLET_TRY_UPGRADE_VAULT"
