@@ -1,18 +1,30 @@
 import "./App.scss";
 import { getAllRouter as AllRouter } from "./router";
-import IdleTimer from "react-idle-timer";
+import { useIdleTimer } from "react-idle-timer";
 import { sendMsg } from "../utils/commonMsg";
 import { WALLET_RESET_LAST_ACTIVE_TIME } from "../constant/msgTypes";
 import { useCallback, useEffect, useState } from "react";
 import cls from "classnames";
-import { fetchSupportTokenInfo, getRecommendFee, getScamList } from "../background/api";
+import {
+  fetchSupportTokenInfo,
+  getRecommendFee,
+  getScamList,
+} from "../background/api";
 import { updateRecommendFee } from "../reducers/cache";
 import { useDispatch, useSelector } from "react-redux";
 import { getLocal } from "../background/localStorage";
-import { RECOMMEND_FEE, SCAM_LIST, SUPPORT_TOKEN_LIST } from "../constant/storageKey";
-import { updateScamList, updateSupportTokenList } from "../reducers/accountReducer";
+import {
+  RECOMMEND_FEE,
+  SCAM_LIST,
+  SUPPORT_TOKEN_LIST,
+} from "../constant/storageKey";
+import {
+  updateScamList,
+  updateSupportTokenList,
+} from "../reducers/accountReducer";
 import { NetworkID_MAP } from "@/constant/network";
 import { getReadableNetworkId } from "@/utils/utils";
+import LedgerStatusSyncer from "./component/LedgerStatusSyncer";
 
 function setLastActiveTime() {
   sendMsg(
@@ -32,26 +44,23 @@ function App() {
   const [autoWidthStatus, setAutoWidthStatus] = useState(false);
   useEffect(() => {
     const url = new URL(window.location.href);
-    const ledgerPageList = ["popup.html#/ledger_page"];
-    const initPageList = [
-      "popup.html#/register_page",
-      "popup.html#/createprocess",
-    ];
-    const zkPage = ["popup.html#/approve_page", "popup.html#/request_sign","popup.html#/token_sign"];
+    // Extract hash path without query parameters (e.g., "#/register_page?addWallet=true" -> "/register_page")
+    const hashPath = url.hash.split("?")[0].replace("#", "");
 
-    zkPage.map((path) => {
-      if (url.href.indexOf(path) !== -1) {
-        setAutoWidthStatus(true);
-      }
-    });
-    let findIndex = false;
-    [...ledgerPageList, ...initPageList].map((path) => {
-      if (url.href.indexOf(path) !== -1) {
-        findIndex = true;
-      }
-    });
+    const fullPageRoutes = ["/ledger_page", "/register_page", "/createprocess"];
+    const zkPageRoutes = ["/approve_page", "/request_sign", "/token_sign"];
+
+    const isZkPage = zkPageRoutes.some((route) => hashPath.startsWith(route));
+    const isFullPage = fullPageRoutes.some((route) =>
+      hashPath.startsWith(route)
+    );
+
+    if (isZkPage) {
+      setAutoWidthStatus(true);
+    }
+
     if (
-      (url.pathname.indexOf("popup.html") !== -1 && !findIndex) ||
+      (url.pathname.indexOf("popup.html") !== -1 && !isFullPage) ||
       url.pathname.indexOf("notification.html") !== -1
     ) {
       setShowFullStatus(false);
@@ -89,12 +98,12 @@ function App() {
     }
   }, []);
 
-  const initTokenNetInfo = useCallback(async()=>{
+  const initTokenNetInfo = useCallback(async () => {
     let tokenInfoList = await fetchSupportTokenInfo();
     if (tokenInfoList.length > 0) {
       dispatch(updateSupportTokenList(tokenInfoList));
     }
-  },[])
+  }, []);
 
   const initNetData = useCallback(() => {
     fetchFeeData();
@@ -115,36 +124,41 @@ function App() {
     initNetData();
   }, []);
 
-  const loadLocalSupportToken = useCallback(()=>{
+  const loadLocalSupportToken = useCallback(() => {
     const readableNetworkId = getReadableNetworkId(currentNode.networkID);
-   let supportList = getLocal(SUPPORT_TOKEN_LIST + "_" + readableNetworkId)
-   if (supportList) {
-     let supportListJson = JSON.parse(supportList)
-     if (supportListJson) {
-       dispatch(updateSupportTokenList(supportListJson));
-     }
-   }
- },[currentNode])
+    let supportList = getLocal(SUPPORT_TOKEN_LIST + "_" + readableNetworkId);
+    if (supportList) {
+      let supportListJson = JSON.parse(supportList);
+      if (supportListJson) {
+        dispatch(updateSupportTokenList(supportListJson));
+      }
+    }
+  }, [currentNode]);
 
   useEffect(() => {
-    if(shouldRefresh && currentNode?.networkID){
-      loadLocalSupportToken()
+    if (shouldRefresh && currentNode?.networkID) {
+      loadLocalSupportToken();
       initTokenNetInfo();
     }
-  }, [shouldRefresh,currentNode]);
+  }, [shouldRefresh, currentNode]);
+
+  useIdleTimer({
+    onAction: setLastActiveTime,
+    throttle: 1000,
+    timeout: 1000 * 60 * 30,
+  });
 
   return (
     <div className="App">
-      <IdleTimer onAction={setLastActiveTime} throttle={1000}>
-        <header
-          className={cls("App-header", {
-            "App-header-full": showFullStatus,
-            AppAutoWidth: autoWidthStatus,
-          })}
-        >
-          <AllRouter />
-        </header>
-      </IdleTimer>
+      <LedgerStatusSyncer />
+      <header
+        className={cls("App-header", {
+          "App-header-full": showFullStatus,
+          AppAutoWidth: autoWidthStatus,
+        })}
+      >
+        <AllRouter />
+      </header>
     </div>
   );
 }
