@@ -1,27 +1,14 @@
 import { WALLET_CREATE_PWD } from "@/constant/msgTypes";
 import Button from "@/popup/component/Button";
 import Input from "@/popup/component/Input";
+import ProcessLayout from "@/popup/component/ProcessLayout";
 import { ReminderTip } from "@/popup/component/ReminderTip";
 import { sendMsg } from "@/utils/commonMsg";
 import i18n from "i18next";
 import { useCallback, useState } from "react";
-import { useHistory } from "react-router-dom";
 import styled from "styled-components";
-import { BackView } from ".";
 import { PasswordValidationList } from "../../../utils/utils";
 
-const StyledPwdContainer = styled.div`
-  margin-top: 20px;
-`;
-const StyledPwdContentContainer = styled.div`
-  padding: 40px;
-`;
-const StyledProcessTitle = styled.div`
-  color: var(--Black, #000);
-  font-size: 24px;
-  font-weight: 700;
-  margin-bottom: 40px;
-`;
 const StyledPwdInputContainer = styled.div`
   margin-top: 20px;
   max-width: 335px;
@@ -37,13 +24,6 @@ const StyledPwdCheckSpan = styled.span`
   color: var(--secondaryRed);
   display: ${(props) => (props.showstatus == "true" ? "none" : "initial")};
 `;
-const StyledBottomContainer = styled.div`
-  position: absolute;
-  width: 600px;
-  bottom: 30px;
-  left: 50%;
-  transform: translate(-50%);
-`;
 export const CreatePwdView = ({ onClickNextTab, onClickPre }) => {
   const [inputPwd, setInputPwd] = useState("");
 
@@ -54,7 +34,6 @@ export const CreatePwdView = ({ onClickNextTab, onClickPre }) => {
   const [confirmPwd, setConfirmPwd] = useState("");
   const [passwordsMatch, setPasswordsMatch] = useState(true);
 
-  const history = useHistory();
 
   const onPwdInput = useCallback(
     (e) => {
@@ -84,9 +63,6 @@ export const CreatePwdView = ({ onClickNextTab, onClickPre }) => {
     [inputPwd]
   );
 
-  const onClickBack = useCallback(() => {
-    history.goBack();
-  }, []);
 
   const goToCreate = useCallback(() => {
     sendMsg({
@@ -122,41 +98,40 @@ export const CreatePwdView = ({ onClickNextTab, onClickPre }) => {
     return txt;
   };
   return (
-    <StyledPwdContainer>
-      <BackView onClickBack={onClickPre} />
-      <StyledPwdContentContainer>
-        <StyledProcessTitle>{i18n.t("createPassword")}</StyledProcessTitle>
-        <ReminderTip content={i18n.t("createPasswordTip")} />
-        <StyledPwdInputContainer>
-          <Input
-            label={i18n.t("password")}
-            onChange={onPwdInput}
-            value={inputPwd}
-            inputType={"password"}
-            showBottomTip={inputPwd.length > 0}
-            bottomTip={
-              <StyledPwdCheckSpan>{getFeedbackText()}</StyledPwdCheckSpan>
-            }
-          />
-          <Input
-            label={i18n.t("confirmPassword")}
-            onChange={onPwdConfirmInput}
-            value={confirmPwd}
-            inputType={"password"}
-            showBottomTip={!passwordsMatch}
-            bottomTip={
-              <StyledPwdCheckSpan>
-                {getConfirmFeedbackText()}
-              </StyledPwdCheckSpan>
-            }
-          />
-        </StyledPwdInputContainer>
-      </StyledPwdContentContainer>
-      <StyledBottomContainer>
+    <ProcessLayout
+      onClickBack={onClickPre}
+      title={i18n.t("createPassword")}
+      bottomContent={
         <Button disable={!isButtonEnabled()} onClick={goToCreate}>
           {i18n.t("next")}
         </Button>
-      </StyledBottomContainer>
-    </StyledPwdContainer>
+      }
+    >
+      <ReminderTip content={i18n.t("createPasswordTip")} />
+      <StyledPwdInputContainer>
+        <Input
+          label={i18n.t("password")}
+          onChange={onPwdInput}
+          value={inputPwd}
+          inputType={"password"}
+          showBottomTip={inputPwd.length > 0}
+          bottomTip={
+            <StyledPwdCheckSpan>{getFeedbackText()}</StyledPwdCheckSpan>
+          }
+        />
+        <Input
+          label={i18n.t("confirmPassword")}
+          onChange={onPwdConfirmInput}
+          value={confirmPwd}
+          inputType={"password"}
+          showBottomTip={!passwordsMatch}
+          bottomTip={
+            <StyledPwdCheckSpan>
+              {getConfirmFeedbackText()}
+            </StyledPwdCheckSpan>
+          }
+        />
+      </StyledPwdInputContainer>
+    </ProcessLayout>
   );
 };
