@@ -2,8 +2,10 @@
  * the init wallet process , create pwd , show mne , backup mne , success
  */
 import { WALLET_CREATE_TYPE } from "@/constant/commonType";
-import Tabs, { TAB_TYPE } from "@/popup/component/Tabs";
+import { GET_WALLET_LOCK_STATUS } from "@/constant/msgTypes";
+import StepTabs from "@/popup/component/StepTabs";
 import { StyledPageInnerContent } from "@/popup/style/common";
+import { sendMsg } from "@/utils/commonMsg";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { useSelector } from "react-redux";
 import styled from "styled-components";
@@ -16,18 +18,39 @@ import { RestoreMneView } from "./RestoreMneView";
 const StyledTabContent = styled.div`
   width: 100%;
   height: 100%;
+  flex: 1;
+  display: flex;
+  flex-direction: column;
 `;
 
 export const CreateProcessPage = ({onClickPre}) => {
   const welcomeNextType = useSelector((state) => state.cache.welcomeNextType);
   const [tabIndex, setTabIndex] = useState(0);
-  const [showMore,setShowMore] = useState(false)
+  const [showMore,setShowMore] = useState(false);
+  const [hasExistingWallet, setHasExistingWallet] = useState(false);
+  
+  useEffect(() => {
+    // Check if wallet already exists (user is unlocked means wallet exists)
+    sendMsg({ action: GET_WALLET_LOCK_STATUS }, (isUnlocked) => {
+      if (isUnlocked) {
+        setHasExistingWallet(true);
+        // Skip password creation step for existing wallets
+        setTabIndex(1);
+      }
+    });
+  }, []);
+  
   const onClickNextTab = useCallback(() => {
     setTabIndex((state) => state + 1);
   }, []);
   const onClickPreTab = useCallback(() => {
+    // For existing wallets, don't go back to password step
+    if (hasExistingWallet && tabIndex === 1) {
+      onClickPre && onClickPre();
+      return;
+    }
     setTabIndex((state) => state - 1);
-  }, []);
+  }, [hasExistingWallet, tabIndex, onClickPre]);
 
   const showRestore = useMemo(() => {
     return welcomeNextType === WALLET_CREATE_TYPE.restore;
@@ -45,7 +68,7 @@ export const CreateProcessPage = ({onClickPre}) => {
 
   return (
     <StyledPageInnerContent showMore={showMore}>
-      <Tabs selected={tabIndex} tabType={TAB_TYPE.STEP}>
+      <StepTabs selected={tabIndex}>
         <StyledTabContent id={1}>
           <CreatePwdView onClickNextTab={onClickNextTab} onClickPre={onClickPre}/>
         </StyledTabContent>
@@ -77,26 +100,7 @@ export const CreateProcessPage = ({onClickPre}) => {
         <StyledTabContent id={5}>
           <CreateResultView />
         </StyledTabContent>
-      </Tabs>
+      </StepTabs>
     </StyledPageInnerContent>
-  );
-};
-
-const StyledBackContainer = styled.div`
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  width: fit-content;
-  margin-left: 20px;
-`;
-const StyledBackImg = styled.img`
-  width: 30px;
-  height: 30px;
-`;
-export const BackView = ({ onClickBack }) => {
-  return (
-    <StyledBackContainer onClick={onClickBack}>
-      <StyledBackImg src="/img/icon_back.svg" />
-    </StyledBackContainer>
   );
 };
