@@ -1,9 +1,9 @@
-import { configureStore, getDefaultMiddleware } from "@reduxjs/toolkit";
+import { configureStore } from "@reduxjs/toolkit";
 import rootReducer from '../reducers';
 
 const store = configureStore({
-      reducer: rootReducer,
-      middleware: [...getDefaultMiddleware(),
-      ],
-    });
+  reducer: rootReducer,
+  middleware: (getDefault) => getDefault(),
+});
+
 export default store;
