@@ -1,31 +1,57 @@
 const webpack = require("webpack");
 const path = require("path");
+const fs = require("fs");
 const CopyWebpackPlugin = require("copy-webpack-plugin");
 const TerserPlugin = require("terser-webpack-plugin");
 
 module.exports = (env, argv) => {
-  console.log("argv.mode", argv.mode);
-  const mode = argv.mode;
+  console.log("argv.mode:", argv.mode);
+  const mode = argv.mode || "production";
   const isDev = mode === "development";
+  const browser = env.BROWSER || "chrome";
 
-  const cssRegex = /\.css$/;
-  const cssModuleRegex = /\.module\.css$/;
-  const sassRegex = /\.s[ac]ss$/i;
-  const sassModuleRegex = /\.module\.s[ac]ss$/;
+  console.log("BROWSER:", env.BROWSER);
+  if (browser === "firefox") {
+    const mainFile = path.resolve(__dirname, "src/utils/o1jsUtils.js");
+    const ffFile = path.resolve(__dirname, "src/utils/o1jsUtils.firefox.js");
+
+    if (fs.existsSync(mainFile) && fs.existsSync(ffFile)) {
+      const mainContent = fs.readFileSync(mainFile, "utf-8");
+      const ffContent = fs.readFileSync(ffFile, "utf-8");
+
+      const mainExports = [
+        ...mainContent.matchAll(/export\s+(?:const|function)\s+(\w+)/g),
+      ].map((m) => m[1]);
+      const ffExports = [
+        ...ffContent.matchAll(/export\s+(?:const|function)\s+(\w+)/g),
+      ].map((m) => m[1]);
+
+      const missing = mainExports.filter((name) => !ffExports.includes(name));
+
+      if (missing.length === 0) {
+        console.log("o1jsUtils.firefox.js is cover all methods.");
+      } else {
+        console.warn(
+          "o1jsUtils.firefox.js less these methods:",
+          missing.join(", ")
+        );
+        console.warn("please add, or Firefox will load error!");
+      }
+    } else if (fs.existsSync(mainFile)) {
+      console.warn("Warn: can not find o1jsUtils.firefox.js!");
+    }
+  }
 
   const config = {
-    devtool: false,
+    mode,
+    devtool: isDev ? "cheap-module-source-map" : false,
     optimization: {
-      minimizer: [
-        new TerserPlugin({
-          extractComments: false,
-        }),
-      ],
+      minimizer: [new TerserPlugin({ extractComments: false })],
       splitChunks: {
         chunks: "async",
         minSize: 20000,
         minRemainingSize: 0,
-        maxSize: 0,
+        maxSize: 250000,
         minChunks: 1,
         maxAsyncRequests: 30,
         maxInitialRequests: 30,
@@ -43,6 +69,7 @@ module.exports = (env, argv) => {
           },
         },
       },
+      minimize: !isDev,
     },
     entry: {
       background: [
@@ -52,7 +79,10 @@ module.exports = (env, argv) => {
       popup: "./src/index.js",
       contentScript: "./src/contentScript/index.js",
       webhook: "./src/webHook/index.js",
-      sandbox: "./src/sandbox/index.js",
+      sandbox:
+        browser === "firefox"
+          ? "./src/sandbox/index.firefox.js"
+          : "./src/sandbox/index.js",
     },
     output: {
       path: path.resolve(__dirname, "./dist"),
@@ -61,66 +91,46 @@ module.exports = (env, argv) => {
     module: {
       rules: [
         {
-          test: cssRegex,
-          exclude: cssModuleRegex,
+          test: /\.css$/,
+          exclude: /\.module\.css$/,
           use: [
-            {
-              loader: "style-loader",
-            },
-            {
-              loader: "css-loader",
-              options: {
-                url: false,
-                sourceMap: true,
-              },
-            },
+            "style-loader",
+            { loader: "css-loader", options: { url: false, sourceMap: isDev } },
           ],
         },
         {
-          test: cssModuleRegex,
-          use: [
-            {
-              loader: "css-loader",
-              options: {
-                modules: true,
-              },
-            },
-          ],
+          test: /\.module\.css$/,
+          use: [{ loader: "css-loader", options: { modules: true } }],
         },
         {
           test: /\.(js|jsx)$/,
           exclude: /node_modules/,
-          loader: "babel-loader",
+          use: {
+            loader: "babel-loader",
+            options: {
+              cacheDirectory: true,
+              presets: [
+                ["@babel/preset-env", { targets: "defaults" }],
+                ["@babel/preset-react", { runtime: "automatic" }],
+              ],
+              plugins: [
+                "@babel/plugin-transform-class-properties",
+                ["@babel/plugin-transform-runtime", { regenerator: true }],
+              ],
+            },
+          },
         },
         {
-          test: sassRegex,
-          exclude: sassModuleRegex,
-          use: [
-            {
-              loader: "style-loader",
-            },
-            {
-              loader: "css-loader",
-            },
-            {
-              loader: "sass-loader",
-              options: {
-                sassOptions: {
-                  modules: true,
-                },
-              },
-            },
-          ],
+          test: /\.s[ac]ss$/i,
+          exclude: /\.module\.s[ac]ss$/,
+          use: ["style-loader", "css-loader", "sass-loader"],
         },
         {
-          test: sassModuleRegex,
+          test: /\.module\.s[ac]ss$/,
           use: [
-            {
-              loader: "style-loader",
-            },
+            "style-loader",
             {
               loader: "css-loader",
-              // https://github.com/webpack-contrib/css-loader/blob/master/CHANGELOG.md#700-2024-04-04
               options: {
                 modules: {
                   namedExport: false,
@@ -128,28 +138,21 @@ module.exports = (env, argv) => {
                 },
               },
             },
-            {
-              loader: "sass-loader",
-              options: {
-                sassOptions: {
-                  modules: true,
-                },
-              },
-            },
+            "sass-loader",
           ],
         },
         {
-          test: /\.(png|jp(e*)g|svg|gif)$/,
+          test: /\.(png|jpe?g|svg|gif)$/i,
           type: "asset/resource",
         },
       ],
     },
-    plugins: getPlugins(),
+    plugins: getPlugins(browser),
     performance: getPerformance(),
     resolve: {
       fallback: {
         child_process: "empty",
-        crypto: require.resolve("crypto-browserify"),
+        crypto: false,
         stream: require.resolve("stream-browserify"),
         path: require.resolve("path-browserify"),
         buffer: require.resolve("buffer"),
@@ -157,43 +160,54 @@ module.exports = (env, argv) => {
       alias: {
         "@": path.resolve(__dirname, "src"),
       },
+      extensions: [".js", ".jsx", ".json"],
     },
-    // experiments: {
-    //   topLevelAwait: true,
-    // },
   };
-  if (isDev) {
-    config.devtool = "cheap-module-source-map";
-    config.optimization = {
-      minimize: false,
-    };
-  }
+
   return config;
 };
 
 function getPerformance() {
   return {
     hints: "warning",
-    maxAssetSize: 4 * 1024 * 1024, // 4MB
-    maxEntrypointSize: 4 * 1024 * 1024, // 4MB
+    maxAssetSize: 4 * 1024 * 1024,
+    maxEntrypointSize: 4 * 1024 * 1024,
   };
 }
-function getPlugins() {
-  let plugins = [];
-  plugins.push(
+
+function getPlugins(browser) {
+  const plugins = [
     new CopyWebpackPlugin({
       patterns: [
         { from: "./public/static", to: "./" },
         { from: "./public/img", to: "./img" },
-        // { from: "./public/manifest.json", to: "./" },
         { from: "./src/_locales", to: "./_locales" },
+        { from: "./public/static/popup.html", to: "popup.html" },
       ],
     }),
     new webpack.ProvidePlugin({
       React: "react",
       Buffer: ["buffer", "Buffer"],
       process: "process/browser.js",
-    })
-  );
+    }),
+    new webpack.DefinePlugin({
+      "process.env.BROWSER": JSON.stringify(browser),
+    }),
+    new webpack.NormalModuleReplacementPlugin(
+      /[/\\]o1jsUtils(\.js)?$/,
+      (resource) => {
+        const replacement =
+          browser === "firefox"
+            ? path.resolve(__dirname, "src/utils/o1jsUtils.firefox.js")
+            : path.resolve(__dirname, "src/utils/o1jsUtils.js");
+
+        resource.request = replacement;
+        resource.context = path.dirname(replacement);
+        if (resource.createData) {
+          resource.createData.resource = replacement;
+        }
+      }
+    ),
+  ];
   return plugins;
 }
